Make application more deterministic
Following "components" were not using random_state from run.py:
- classifier
- oversample

This lead to more changing results between multiple runs.

Unfortunately, it was not possible to make it fully deterministic, maybe due to the floating point representation or the word2vec training or other unknown reasons. The following things were also tried (all at the same time with the above changes):
- trying to use the same seed in create_w2v_model() from model.py
- no stratifying (stratify=None) and no shuffling (shuffle=False) in train_validation_test_split() from tweet_sentiment_analyzer.py.
- not equally distributing the class labels (by applying even-odd technique: 0-1-0-1, ...) in train (with distribute_labels_equally_in_train())
However, the last three changes were not applied to this commit, because making sure class labels are stratified and equally distributed were assumed to be useful.

Added random_state instance variable to tweet_sentiment_analyzer.py to use it for every called method that makes use of it.